REFACTOR: Moved video url form in edit to a separate component
