+++ resources/ts/Pages/Events/Step/Past/Edit.tsx
@@ -7,6 +7,7 @@ import InputLabel2 from "@/Components/Forms/InputLabel2";
 import RadioInput from "@/Components/Forms/RadioInput";
 import TextInput from "@/Components/Forms/TextInput";
 import ToastBanner from "@/Components/Forms/ToastBanner";
+import VideoEditFormComponent from "@/Components/VideoEditFormComponent";
 import ContentWrapper from "@/Layouts/ContentWrapper";
 import WrapperLayout from "@/Layouts/WrapperLayout";
 import { usePage, useForm } from "@inertiajs/react";
@@ -22,66 +23,22 @@ interface FullVideoMeta {
     imageFile?: File | null,
     imageLink: string,
     content: VideoMeta[],
+    fileContent: FileMeta[],
     showDetails: "1" | "0";
 }
 
 
-export default function Edit({ videoData }: { videoData: FullVideoMeta }) {
-
-    interface Action {
-        type: "changeValue" | "addValue" | "removeValue";
-    }
-    interface ChangeAction extends Action {
-        name: keyof VideoMeta;
-        value: string | number;
-        idx: number;
-    }
-    const initialState: VideoMeta[] = videoData.content;
-    const blankState: VideoMeta[] = [{
-        title: "",
-        externalUrl: "",
-        duration: "",
-        id: 0
-    }];
-
-    const reducer = (state: VideoMeta[], action: ChangeAction | Action) => {
-        if (action.type === "changeValue" && "name" in action) {
-            let returnObj = [...state];
-            if (action.name === "id") {
-                returnObj[action.idx][action.name] = +action.value;
-            } else {
-                returnObj[action.idx][action.name] = action.value + "";
-            }
-            return returnObj;
-        } else if (action.type === 'addValue') {
-            return [
-                ...state, ...blankState
-            ];
-        } else if (action.type === "removeValue" && "idx" in action) {
-            if (state.length === 1) {
-                return blankState;
-            }
-            let returnObj = [...state];
-            returnObj.splice(action.idx, 1);
-            return returnObj;
-        }
-        else {
-            return state;
-        }
-    }
-
-    const [videoState, dispatch] = useReducer(reducer, initialState);
-
+export default function Edit({ videoData: eventData }: { videoData: FullVideoMeta }) {
 
     const { errors } = usePage().props;
     const { data, setData, post, reset, processing } = useForm({
-        date: videoData.date,
-        heading: videoData.heading,
-        description: videoData.description,
-        imageFile: videoData.imageFile,
-        imageLink: videoData.imageLink,
-        showDetails: videoData.showDetails,
-        content: videoData.content
+        date: eventData.date,
+        heading: eventData.heading,
+        description: eventData.description,
+        imageFile: eventData.imageFile,
+        imageLink: eventData.imageLink,
+        showDetails: eventData.showDetails,
+        content: eventData.content
     });
 
     const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -98,36 +55,19 @@ export default function Edit({ videoData }: { videoData: FullVideoMeta }) {
             setData(event.target.name, event.target.files[0]);
         }
     }
-    const handleComplexChange = (idx: number, event: React.ChangeEvent<HTMLInputElement | HTMLElement>) => {
-        if (event.target instanceof HTMLInputElement) {
-            switch (event.target.name) {
-                case "title":
-                case "externalUrl":
-                case "duration":
-                case "id":
-                    dispatch({
-                        type: "changeValue",
-                        name: event.target.name,
-                        value: event.target.value,
-                        idx: idx
-                    });
-                    break;
-            }
-        }
-    }
 
     const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
         event.preventDefault();
-        post(route('events.step.past.update', videoData.id));
+        post(route('events.step.past.update', eventData.id));
+    }
+
+    const updateContent = (newContent: VideoMeta[]) => {
+        setData('content', [...newContent]);
     }
 
     useEffect(() => {
         reset();
     }, []);
-
-    useEffect(() => {
-        setData("content", [...videoState]);
-    }, [videoState]);
 
 
     return (
@@ -172,76 +112,7 @@ export default function Edit({ videoData }: { videoData: FullVideoMeta }) {
                         </div>
                         <img className="w-60" src={data.imageFile ? URL.createObjectURL(data.imageFile) : data.imageLink} />
                     </div>
-                    <h2 className="p-0 mb-2 text-xl font-bold text-black">Video Information</h2>
-                    <p className="text-gray-600">Change the ID number to change order of video. Be careful of the maximum number of videos</p>
-                    <table>
-                        <thead>
-                            <tr>
-                                <th>ID</th>
-                                <th>External URL</th>
-                                <th>Title</th>
-                                <th>Duration</th>
-                                <th><SecondaryButton onClick={() => dispatch({ type: "addValue" })} className="before:content-['+'] before:pr-1 before:text-lg bg-green-200">Add Row</SecondaryButton></th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {videoState.map(({ title, externalUrl, duration, id }, idx) => (
-                                <tr key={"contenttable" + idx}>
-                                    <td>
-                                        <TextInput
-                                            type={"text"}
-                                            name={"id"}
-                                            id={`id${idx}`}
-                                            value={id}
-                                            className={""}
-                                            handleChange={(e) => handleComplexChange(idx, e)}
-                                            required
-                                        />
-                                    </td>
-                                    <td>
-                                        <TextInput
-                                            type={"text"}
-                                            name={"externalUrl"}
-                                            id={`externalUrl${idx}`}
-                                            value={externalUrl}
-                                            className={""}
-                                            handleChange={(e) => handleComplexChange(idx, e)}
-                                        />
-                                    </td>
-                                    <td>
-                                        <TextInput
-                                            type={"text"}
-                                            name={"title"}
-                                            id={`title${idx}`}
-                                            value={title}
-                                            className={""}
-                                            handleChange={(e) => handleComplexChange(idx, e)}
-                                        />
-                                    </td>
-                                    <td>
-                                        <TextInput
-                                            type={"text"}
-                                            name={"duration"}
-                                            id={`duration${idx}`}
-                                            value={duration}
-                                            className={""}
-                                            handleChange={(e) => handleComplexChange(idx, e)}
-                                        />
-                                    </td>
-                                    <td>
-                                        <SecondaryButton
-                                            onClick={() => dispatch({
-                                                type: "removeValue",
-                                                idx: idx
-                                            })}
-                                            className="bg-red-200 before:content-['-'] before:pr-1 before:text-lg">
-                                            Remove Row
-                                        </SecondaryButton>
-                                    </td>
-                                </tr>
-                            ))}
-                        </tbody>
-                    </table>
+                    <VideoEditFormComponent videoContent={eventData.content} updateContent={updateContent} />
 
                     <div className="inline-flex justify-center w-full gap-2 mt-5 md:justify-end">
                         <ButtonLink type="secondary" href={route('events.step.past.admin')}>Cancel</ButtonLink>
